Remove the parseBreakends by default functionality
